refactoring: responsible classes to different modules
